chore: move stack trace to its own entry point
