+++ packages/utils/src/index.ts
@@ -7,4 +7,4 @@ export * from './display'
 export * from './constants'
 export * from './colors'
 export * from './base'
-export * from './source-map'
+export * from './offset'
