+++ packages/vitest/src/api/setup.ts
@@ -7,7 +7,7 @@ import { parse, stringify } from 'flatted'
 import type { WebSocket } from 'ws'
 import { WebSocketServer } from 'ws'
 import type { ViteDevServer } from 'vite'
-import type { StackTraceParserOptions } from '@vitest/utils'
+import type { StackTraceParserOptions } from '@vitest/utils/source-map'
 import { API_PATH } from '../constants'
 import type { Vitest } from '../node'
 import type { File, ModuleGraphData, Reporter, TaskResultPack, UserConsoleLog } from '../types'
