+++ packages/vitest/src/typecheck/typechecker.ts
@@ -2,7 +2,7 @@ import { rm } from 'node:fs/promises'
 import type { ExecaChildProcess } from 'execa'
 import { execa } from 'execa'
 import { basename, extname, resolve } from 'pathe'
-import { TraceMap, generatedPositionFor } from '@jridgewell/trace-mapping'
+import { TraceMap, generatedPositionFor } from '@vitest/utils/source-map'
 import type { RawSourceMap } from '@ampproject/remapping'
 import { getTasks } from '../utils'
 import { ensurePackageInstalled } from '../node/pkg'
