+++ packages/vitest/src/utils/source-map.ts
@@ -1,8 +1,10 @@
 export {
   lineSplitRE,
-  parseSingleStack,
-  parseStacktrace,
-  parseErrorStacktrace,
   positionToOffset,
   offsetToLineNumber,
 } from '@vitest/utils'
+export {
+  parseSingleStack,
+  parseStacktrace,
+  parseErrorStacktrace,
+} from '@vitest/utils/source-map'
